OEL-2266: Rename also classes.
