+++ tests/src/Traits/BackwardCompatibilityTrait.php
@@ -4,10 +4,12 @@ declare(strict_types = 1);
 
 namespace Drupal\Tests\oe_bootstrap_theme\Traits;
 
+use Drupal\oe_bootstrap_theme\BackwardCompatibility;
+
 /**
  * Contains method that handle backwards compatibility.
  */
-trait BackwardsCompatibilityTrait {
+trait BackwardCompatibilityTrait {
 
   /**
    * Sets a backward compatibility setting.
@@ -19,7 +21,7 @@ trait BackwardsCompatibilityTrait {
    */
   protected function setBackwardCompatibilitySetting(string $name, bool $value): void {
     $theme_config = $this->container->get('config.factory')->getEditable('oe_bootstrap_theme.settings');
-    $theme_config->set('backward_compatibility.' . $name, $value)->save();
+    $theme_config->set(BackwardCompatibility::PREFIX . $name, $value)->save();
     drupal_static_reset();
   }
